requirements.txt
Pillow==2.8.2
wheel==0.24.0
